Implement file upload in two stages (init+upload)
Firs stage initializes the upload. It sends a post requst stating that
the upload will take place shortly. It's a post request, for it creates
a UploadedTask model instance.

Second stage performs the actual upload passing the created UploadedTask
instance id as a parameter. Thus, the system is able to save progress
data related to this UploadedTask instance.

The progress data may later be obtained using same pk.